Rifle audio fix!
Put in some code that checked if the rifle has the semi-auto upgrade installed. If it is, it plays the original rifle shooting audio (still with the distance effects).

Otherwise, It will play a new audio which includes the bolt action after the shot. This can only be heard in the tile of the player, otherwise It will use the same sound as the semi-auto shot in the distance