WP update
expanded cloud, made WP "shrapnel" for entity interaction. Shrapnel needs spread, gravity and to generate WP cloud on ground impact, along with fire

NOTE:
Looking at the old system, WP was treated as a bullet entity, which is no longer in usage. This isnt as much of a bug as it it an implementation.